+++ model/src/main/java/org/software/model/content/vo/CommentVO.java
@@ -7,12 +7,12 @@ import java.util.List;
 
 @Data
 public class CommentVO {
-    private Integer commentId;
-    private Integer contentId;
-    private Integer userId;
-    private Integer parentCommentId;
-    private Integer rootCommentId;
-    private Integer toUserId;
+    private Long commentId;
+    private Long contentId;
+    private Long userId;
+    private Long parentCommentId;
+    private Long rootCommentId;
+    private Long toUserId;
     private String content;
     private Integer isRead;
     private Date createdAt;
