+++ model/src/main/java/org/software/model/content/vo/ContentLikeFavoriteVO.java
@@ -6,9 +6,9 @@ import java.util.Date;
 
 @Data
 public class ContentLikeFavoriteVO {
-    private Integer likeId;
-    private Integer contentId;
-    private Integer userId;
+    private Long likeId;
+    private Long contentId;
+    private Long userId;
     private Integer isRead;
     private String type;
     private Date createdAt;
