+++ service/service-content/src/main/java/org/software/content/controller/CommentController.java
@@ -22,29 +22,29 @@ public class CommentController {
     // 新增评论
     @PostMapping
     public Response addComment(@RequestBody CommentDTO commentDTO) {
-        Integer commentId = commentsService.addComment(commentDTO);
+        Long commentId = commentsService.addComment(commentDTO);
         return Response.success(commentId);
     }
 
     // 获取根评论
     @GetMapping("/root")
-    public Response getRootComments(@RequestParam Integer contentId) {
+    public Response getRootComments(@RequestParam Long contentId) {
         List<CommentVO> rootComments = commentsService.getRootComments(contentId);
         return Response.success(rootComments);
     }
     @GetMapping("/child")
-    public Response getChildComments(@RequestParam Integer commentId) {
+    public Response getChildComments(@RequestParam Long commentId) {
         List<CommentVO> childComments = commentsService.getChildComments(commentId);
         return Response.success(childComments);
     }
     @GetMapping("/unread")
-    public Response getUnreadComments(@RequestParam Integer commentId){
+    public Response getUnreadComments(@RequestParam Long commentId){
         List<CommentVO> unreadComments = commentsService.getUnreadComments(commentId);
         return Response.success(unreadComments);
     }
     @GetMapping("/unread/count")
-    public Response getUnreadCommentsCount(@RequestParam Integer parentCommentId){
-        Integer count=commentsService.getUnreadCommentCount(parentCommentId);
+    public Response getUnreadCommentsCount(@RequestParam Long parentCommentId){
+        Long count=commentsService.getUnreadCommentCount(parentCommentId);
         return Response.success(count);
     }
 }
